The Required Fields and Examples are incorrect
I've updated the examples to be more complete and to have the correct PowerShell switches and parameters. The example included a -web parameter on the Get-SPAppPrincipal command which is incorrect. It also included an incorrect value for the -scope parameter in the Set-SPAppPrincipalPermissions command. 

Also updated the required parameters details as they were incorrect for the -scope and -right parameters.